Complete solver

+++ src/Nono.Engine/Constaints.cs
@@ -2,6 +2,6 @@ namespace Nono.Engine
 {
     public static class Constraints
     {
-        public const uint MIN_BLOCK_SPACE = 1;
+        public const byte MIN_BLOCK_SPACE = 1;
     }
 }
